+++ module02/src/main/java/ru/sberbank/edu/App.java
@@ -1,8 +1,10 @@
 package ru.sberbank.edu;
 
-import ru.sberbank.edu.saver.FileSaver;
+import ru.sberbank.edu.storage.DBStorage;
+import ru.sberbank.edu.storage.FileStorage;
 import ru.sberbank.edu.statistics.RowStatistic;
 import ru.sberbank.edu.statistics.Statistic;
+import ru.sberbank.edu.storage.Storage;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -22,6 +24,11 @@ public class App
         Path statPath = Paths.get("module02", "src","test","resources" , "statistics.txt");
         String absStatPath = statPath.toFile().getAbsolutePath();
 
-        new FileSaver(absStatPath).save(statistic);
+        Storage storage = new FileStorage(absStatPath);
+        storage.save(statistic);
+
+        Storage storage1 = new DBStorage("hana");
+        storage1.save(statistic);
+
     }
 }
