+++ module02/src/main/java/ru/sberbank/edu/storage/Storage.java
@@ -1,9 +1,9 @@
-package ru.sberbank.edu.saver;
+package ru.sberbank.edu.storage;
 
 import ru.sberbank.edu.statistics.Statistic;
 import java.io.IOException;
 
-public interface Saver {
+public interface Storage {
 
     void save(Statistic statistic) throws IOException;
